requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
aniso8601==1.1.0          # via flask-restful
flask-restful==0.3.4
flask==0.10.1
gunicorn==19.3.0
itsdangerous==0.24        # via flask
jinja2==2.8               # via flask
markupsafe==0.23          # via jinja2
pillow==3.0.0
python-dateutil==2.4.2    # via aniso8601
pytz==2015.7              # via flask-restful
six==1.10.0               # via flask-restful, python-dateutil
werkzeug==0.11.1          # via flask
